+++ lib/catcher_plugin.dart
@@ -1,0 +1,22 @@
+library catcher;
+
+export "package:catcher/handlers/console_handler.dart";
+export "package:catcher/handlers/email_auto_handler.dart";
+export "package:catcher/handlers/file_handler.dart";
+export "package:catcher/handlers/http_handler.dart";
+export "package:catcher/handlers/report_handler.dart";
+export "package:catcher/handlers/toast_handler.dart";
+export "package:catcher/handlers/email_manual_handler.dart";
+export "package:catcher/mode/notification_report_mode.dart";
+export 'package:catcher/model/report_mode.dart';
+export 'package:catcher/mode/silent_report_mode.dart';
+export 'package:catcher/mode/dialog_report_mode.dart';
+export 'package:catcher/mode/page_report_mode.dart';
+export 'package:catcher/mode/report_mode_action_confirmed.dart';
+export 'package:catcher/model/http_request_type.dart';
+export 'package:catcher/model/report.dart';
+export 'package:catcher/model/report_mode_type.dart';
+export 'package:catcher/model/toast_handler_gravity.dart';
+export 'package:catcher/model/toast_handler_length.dart';
+export 'package:catcher/model/catcher_options.dart';
+export 'package:catcher/core/catcher.dart';