+++ lib/handlers/report_handler.dart
@@ -1,7 +1,7 @@
 import 'package:catcher/model/report.dart';
 
 abstract class ReportHandler {
-  const ReportHandler();
+  ReportHandler();
 
   Future<bool> handle(Report error);
 }