+++ lib/model/report_mode_type.dart
@@ -1,1 +1,1 @@
-enum ReportModeType { silent, notification }
+enum ReportModeType { silent, notification, page, dialog }