+++ CHANGELOG.md
@@ -1,5 +1,14 @@
+## 0.0.8
+* Added email manual handler
+* Changed email handler to email auto handler
+* Added application profile feature
+* Added page report mode feature
+* Added dialog report mode feature
+* Changed configuration step, it uses now CatcherOptions to setup config for specific profile
+
+
 ## 0.0.7
-Fixed import error
+* Fixed import error
 
 ## 0.0.6
 * Fixed EmailHandler error